Close remaining #1447 gaps, harden the Low preset for old GPUs, fix main's red lanes (#1498)
* Close the remaining #1447 gaps, harden the low preset, fix main's red lanes

Architecture (issue #1447):
- Render: seed Carthaginian elephants from the EntityID, not the Entity*
  address; one GL lifetime policy (textures, shaders, the wear volume and
  meshes release or defer by share group, never call GL without a context);
  meshes re-upload in a foreign share group; atomic GraphicsSettings;
  shader reload skips already-compiled and foreign-group programs; the
  runtime-bake barrier is lifted only by the renderer that raised it; the
  portrait renderer gets its own frame profile; a debug freeze asserts no
  component layout change during parallel prepare.
- Sim: replay digest gains a "systems" part built from opt-in SessionSnapshot
  digest views (AI, undead zones, veins, ground wildlife); replay format 5.
  Victory, mission-wave and wildlife clocks accumulate in double (a float
  running sum drifted ~0.16 s per 1000 s). MapTransformer resolves owners,
  nations, terrain, collision and factories through the world's bound
  services; session_for counts unbound fallbacks. Dropped ticks show on F10.
- Determinism: TerrainService revisions are process-unique, so a second match
  can no longer reuse the previous map's nav grid / prop index (this was why
  AiDuelMatchTest.HannibalMirror failed after another AI match in the same
  binary); the RPG crit roll and FPV swing variant no longer use process-wide
  sequences; TroopProfileService is thread-safe (render thread raced the sim).

Performance (digests bit-identical before/after on 13 scenarios):
- exact squared-distance prefilters in walk_formation_slot and body impact;
  per-tick formation rank roster instead of O(F^2) scans; layout copies
  reuse capacity; troop profiles by reference; gate loop hoisting; cull
  radius memoised; horse mount frame samples only the two joints it needs.
  sim_benchmark 2000 units: ~41 -> ~37 ms/tick interleaved A/B.

Low preset / old GPUs:
- first run picks Low on software renderers and sub-4.3 GL (not on macOS);
  Low uses an R11F_G11F_B10F post target, 4 local lights, smaller terrain
  atlases; bigger palette ring for the per-creature fallback; unsynchronised
  fenced fallback mapping; validator check that 4.x calls stay gated.
  Verified: renders and presents on Mesa llvmpipe capped at GL 3.3 Core.

Fixed on main: RpgFpvOverlay imported its own module (arena HUD failed to
load); 20 uncatalogued overlay strings translated in all eight languages;
UBSan downcast in gl_lifetime; PR lane now runs a short headless replay.


* Format the rendering architecture doc with Prettier


* Drop the last unused tree helper the Clang/libc++ lane rejects


---------

+++ game/mission/mission_wave_runtime.h
@@ -45,7 +45,7 @@ struct MissionWaveBinding {
 
 class MissionWaveRuntime {
 public:
-  [[nodiscard]] auto elapsed() const -> float { return m_elapsed; }
+  [[nodiscard]] auto elapsed() const -> float { return static_cast<float>(m_elapsed); }
   [[nodiscard]] auto director() -> Game::Mission::MissionWaveDirector& {
     return m_director;
   }
@@ -68,7 +68,7 @@ public:
 private:
   [[nodiscard]] auto fire_due_events() -> QStringList;
 
-  float m_elapsed = 0.0F;
+  double m_elapsed = 0.0;
   std::vector<Game::Mission::PendingMissionWave> m_waves;
   std::vector<Game::Mission::PendingMissionEvent> m_events;
   Game::Mission::MissionWaveDirector m_director;
